Development

+++ src/main/java/be/lasse/backend2/fooddelivery/model/Dish.java
@@ -9,9 +9,9 @@ import jakarta.persistence.Id;
 public class Dish {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
     private String description;
-    private int price;
+    private Double price;
 }
